requirements.txt
actionlib==1.12.1
angles==1.9.12
base_local_planner==1.16.7
bondpy==1.8.5
camera_calibration==1.15.0
camera_calibration_parsers==1.11.13
catkin==0.7.29
controller_manager==0.18.4
controller_manager_msgs==0.18.4
cv_bridge==1.13.0
cycler==0.10.0
diagnostic_analysis==1.9.7
diagnostic_common_diagnostics==1.9.7
diagnostic_updater==1.9.7
dynamic_reconfigure==1.6.3
gazebo_plugins==2.8.7
gazebo_ros==2.8.7
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.16
image_geometry==1.13.0
interactive-markers==1.11.5
joblib==1.1.0
joint_state_publisher==1.12.15
kdl_parser_py==1.13.1
kiwisolver==1.3.2
laser_geometry==1.6.7
matplotlib==3.4.3
message_filters==1.14.11
moveit-core==1.0.8
moveit_commander==1.0.8
moveit_ros_planning_interface==1.0.8
moveit_ros_visualization==1.0.8
numpy==1.21.2
openni2_launch==1.5.1
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
python_qt_binding==0.4.4
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
resource_retriever==1.12.6
rosbag==1.14.11
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.11
roslaunch==1.14.11
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.11
rosmake==1.14.9
rosmaster==1.14.11
rosmsg==1.14.11
rosnode==1.14.11
rosparam==1.14.11
rospy==1.14.11
rosserial_python==0.8.0
rosservice==1.14.11
rostest==1.14.11
rostopic==1.14.11
rosunit==1.14.9
roswtf==1.14.11
rqt-moveit==0.5.10
rqt-reconfigure==0.5.4
rqt-robot-monitor==0.5.13
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.9
rqt_dep==0.4.9
rqt_graph==0.4.11
rqt_gui==0.5.2
rqt_gui_py==0.5.2
rqt_image_view==0.4.16
rqt_launch==0.4.8
rqt_logger_level==0.4.8
rqt_msg==0.4.8
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.8
rqt_publisher==0.4.8
rqt_py_common==0.5.2
rqt_py_console==0.4.8
rqt_robot_dashboard==0.5.7
rqt_robot_steering==0.5.10
rqt_runtime_monitor==0.5.7
rqt_service_caller==0.4.8
rqt_shell==0.4.9
rqt_srv==0.4.8
rqt_tf_tree==0.6.0
rqt_top==0.4.8
rqt_topic==0.4.11
rqt_web==0.4.8
rviz==1.13.18
scikit-learn==1.0
scipy==1.7.1
sensor-msgs==1.12.8
six==1.16.0
sklearn==0.0
smach==2.0.1
smach_ros==2.0.1
smclib==1.8.5
srdfdom==0.5.2
tf==1.12.1
tf2_geometry_msgs==0.6.5
tf2_kdl==0.6.5
tf2_py==0.6.5
tf2_ros==0.6.5
tf_conversions==1.12.1
threadpoolctl==3.0.0
topic_tools==1.14.11
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.3
turtlebot3_example==1.2.5
turtlebot3_teleop==1.2.5
typing-extensions==3.10.0.2
urdf_test==1.0.4
urdfdom-py==0.4.5
xacro==1.13.12
